Standardize on a usage of route functions
Always wrap them in either a go-to or make-link

When we make a link we need to root the url with "/#" to avoid it being
sent to the server.  When we want to set the history token, we don't
want it to have any prefix because the history object will handle that.