Fix manga-reader page list integration test
The core issue was that sometimes Manga Fox was serving us a page
encoded with gzip. This looks like nonsense if we're expecting text.
The ultimate answer was to pull in the clj-http library to handle the
http request fully, and to use tempfile to write the html to disk
temporarily so that the incredibly large comic list doesn't do bad
things.