Fix sites.read-test
These tests were modifying the classpath file system and dependent on
the idea that they could write into the classpath at will by putting
things under "resources".

The price of boot classpath isolation is more complicated testing
setup machinery to mess with the test classpath. Upside is the tests
are much more robustly implemented because of the logic extraction.