Expand set-current-location event handler to fetch more images
Don't send an API call if we've already received a response that
indicates that we've reached the end in that direction, but otherwise,
if the buffer of loaded locations in a direction is too small, load
some more.

WIP! start working on issuing api calls for fetching more locations

fixup! WIP! start working on issuing api calls for fetching more locations