Remove site-id from store-comics fn
In this brave new world we're assuming that a comic id is ALWAYS a
keyword qualified by the site id that it belongs to.  So no need. Side
benefit, you can store all the comics you have on hand regardless of
whether they are from the same site.  Dubious value-add for production
code, but useful in the tests.